+++ src/RingGrids/grids/octaminimal_gaussian.jl
@@ -20,7 +20,7 @@ nonparametric_type(::Type{<:OctaminimalGaussianGrid}) = OctaminimalGaussianGrid
 full_grid_type(::Type{<:OctaminimalGaussianGrid}) = FullGaussianGrid
 
 # FIELD
-const OctaminimalGaussianField{T, N} = Field{T, N, Architecture, Grid} where {Architecture, Grid<:OctaminimalGaussianGrid}
+const OctaminimalGaussianField{T, N} = Field{T, N, ArrayType, Grid} where {ArrayType, Grid<:OctaminimalGaussianGrid}
 
 # define grid_type (i) without T, N, (ii) with T, (iii) with T, N but not with <:?Field
 # to not have precendence over grid_type(::Type{Field{...})
@@ -28,8 +28,11 @@ grid_type(::Type{OctaminimalGaussianField}) = OctaminimalGaussianGrid
 grid_type(::Type{OctaminimalGaussianField{T}}) where T = OctaminimalGaussianGrid
 grid_type(::Type{OctaminimalGaussianField{T, N}}) where {T, N} = OctaminimalGaussianGrid
 
-Base.show(io::IO, F::Type{<:OctaminimalGaussianField{T, N}}) where {T, N} =
+function Base.showarg(io::IO, F::Field{T, N, ArrayType, Grid}, toplevel) where {T, N, ArrayType, Grid<:OctaminimalGaussianGrid{A}} where A <: AbstractArchitecture
     print(io, "OctaminimalGaussianField{$T, $N}")
+    toplevel && print(io, " on ", nonparametric_type(ArrayType))
+    toplevel && print(io, " on ", A)
+end
 
 # SIZE
 nlat_odd(::Type{<:OctaminimalGaussianGrid}) = false
@@ -71,7 +74,7 @@ get_quadrature_weights(::Type{<:OctaminimalGaussianGrid}, nlat_half::Integer) = 
 """$(TYPEDSIGNATURES) precompute a `Vector{UnitRange{Int}} to index grid points on
 every ring `j` (elements of the vector) of `Grid` at resolution `nlat_half`.
 See `eachring` and `eachgrid` for efficient looping over grid points."""
-function each_index_in_ring!(   rings::AbstractVector,
+function each_index_in_ring!(   rings,
                                 Grid::Type{<:OctaminimalGaussianGrid},
                                 nlat_half::Integer) # resolution param
 
@@ -93,3 +96,5 @@ function each_index_in_ring!(   rings::AbstractVector,
         rings[j] = index_1st:index_end              # turn into UnitRange
     end
 end
+
+Adapt.@adapt_structure OctaminimalGaussianGrid