+++ test/GPU/kernels_GPU.jl
@@ -1,6 +1,7 @@
-using KernelAbstractions
-import SpeedyWeather: on_architecture, CPU, launch!
-@testset "KernelAbstractions tests" begin 
+import SpeedyWeather: on_architecture, GPU, launch!
+import KernelAbstractions: KernelAbstractions, @kernel 
+
+@testset "KernelAbstractions GPU tests" begin 
 
     # To-Do write tests for each type of dims_type in the kernel launching util, 
     # the tests currently below will be removed when the KA becomes the only one
@@ -21,7 +22,7 @@ import SpeedyWeather: on_architecture, CPU, launch!
         N = 2   # number of matrices
         NF = Float32
 
-        arch = SpeedyWeather.CPU()
+        arch = SpeedyWeather.GPU()
 
         # Create test arrays
         B = on_architecture(arch, rand(LowerTriangularArray{Complex{NF}}, L, M, N))
@@ -32,41 +33,9 @@ import SpeedyWeather: on_architecture, CPU, launch!
 
         # Run the kernel
         launch!(arch, :lmk, size(A), test_lta_kernel!, A, B, C)
-        synchronize(arch)
+        KernelAbstractions.synchronize(arch)
 
         # Verify results
         @test A ≈ expected
     end
-
-    @testset "RingGrids kernel test" begin
-
-        Grid = HEALPixGrid
-        @kernel function test_ringgrids_kernel!(A, B, C)
-            I = @index(Global, Linear)
-            A[I] = B[I] * C[I]
-        end
-
-        # Test parameters
-        nlat_half = 6   # degree
-        nlayers = 5
-        NF = Float32
-
-        arch = SpeedyWeather.CPU()
-
-        # Create test arrays
-        B = on_architecture(arch, rand(Grid{NF}, nlat_half, nlayers))
-        C = on_architecture(arch, rand(Grid{NF}, nlat_half, nlayers))
-        A = similar(B)
-
-        expected = B .* C
-
-        # Run the kernel
-        launch!(arch, :ijk, size(A), test_ringgrids_kernel!, A, B, C)
-        synchronize(arch)
-
-        # Verify results
-        @test A ≈ expected
-
-    end     
-
 end