+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Reworked GPU and utility code structure and added GPU versions of spectral gradients[#727](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/727)
 - Added `animate` to animate a NetCDF output file with GeoMakie [#755](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/755)
 - Makie v0.24 compatibility [#750](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/750)
 - Fixed an error in the LowerTriangularArrays documentation [#749](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/749)